session1/pyproject.toml
[project]
name = "session1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.1",
]
